exceed_print_area_project/requirements.txt
asgiref==3.5.0
Django==3.2.9
djangorestframework==3.13.1
PyMuPDF==1.19.6
pytz==2022.1
sqlparse==0.4.2
